requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
faker
apache-airflow